+++ hc/integrations/shell/tests/test_notify.py
@@ -6,6 +6,7 @@ from unittest.mock import Mock, patch
 
 from django.test.utils import override_settings
 from django.utils.timezone import now
+
 from hc.api.models import Channel, Check, Flip, Notification
 from hc.test import BaseTestCase
 
@@ -79,3 +80,60 @@ class NotifyTestCase(BaseTestCase):
 
         n = Notification.objects.get()
         self.assertEqual(n.error, "Shell commands are not enabled")
+
+    @patch("hc.integrations.shell.transport.os.system")
+    @override_settings(SHELL_ENABLED=True)
+    def test_shell_escapes_name(self, mock_system: Mock) -> None:
+        definition = {"cmd_down": "logger $NAME", "cmd_up": ""}
+        self._setup_data("shell", json.dumps(definition))
+        mock_system.return_value = 0
+
+        samples = {
+            "a&b": "logger 'a&b'",
+            "a'b": """logger 'a'"'"'b'""",
+        }
+
+        for name, escaped_cmd in samples.items():
+            self.check.name = name
+            self.check.save()
+            self.channel.notify(self.flip)
+
+            mock_system.assert_called_with(escaped_cmd)
+
+    @patch("hc.integrations.shell.transport.os.system")
+    @override_settings(SHELL_ENABLED=True)
+    def test_shell_escapes_tags(self, mock_system: Mock) -> None:
+        definition = {"cmd_down": "logger $TAGS", "cmd_up": ""}
+        self._setup_data("shell", json.dumps(definition))
+        mock_system.return_value = 0
+
+        samples = {
+            "a&b": "logger 'a&b'",
+            "a'b": """logger 'a'"'"'b'""",
+        }
+
+        for tags, escaped_cmd in samples.items():
+            self.check.tags = tags
+            self.check.save()
+            self.channel.notify(self.flip)
+
+            mock_system.assert_called_with(escaped_cmd)
+
+    @patch("hc.integrations.shell.transport.os.system")
+    @override_settings(SHELL_ENABLED=True)
+    def test_shell_escapes_tag1(self, mock_system: Mock) -> None:
+        definition = {"cmd_down": "logger $TAG1", "cmd_up": ""}
+        self._setup_data("shell", json.dumps(definition))
+        mock_system.return_value = 0
+
+        samples = {
+            "a&b anothertag": "logger 'a&b'",
+            "a'b anothertag": """logger 'a'"'"'b'""",
+        }
+
+        for tags, escaped_cmd in samples.items():
+            self.check.tags = tags
+            self.check.save()
+            self.channel.notify(self.flip)
+
+            mock_system.assert_called_with(escaped_cmd)