+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Fix sendalerts to avoid crashing on a timed out db connection (#1267)
 - Fix the update_timeout view to also kick off nags
 - Fix the email integration to handle deep attachment stacks in last ping body
+- Fix check name and tag escaping in the Shell integration
 
 ## v4.1 - 2026-03-20